Add support for decrypting V2 CSRs
We can now decrypt CSRs that are encrypted to the now-public EEK. This
allows partners to better test and inspect the output of their HALs
than previously, when we considered the EEK to be sensitive.

This decryption support will be included in VTS and
rkp_factory_extraction_tool in the future to better validate pre-v3
implementations.

Bug: 280620556
Test: libhwtrust_tests
Change-Id: Id14d74c00ff64b67eceb9120c86ae17bcbaf8c4f