Read profile name from DICE chain entries
Parse the declared profile version from the DICE chain entries or try to
infer the profile version from context if it's not explicitly declared.
Select the profile valiadtion rules based on the profile version.

Session options now select the range of DICE profiles to support rather
than the parsing parameters for a specific DICE profile. The options for
a given VSR set the minimum supported profile version for compatibility
with older hardware. The maximum supported DICE version is set to the
version that corresponds with the same Android version as the VSR.

Test: atest --host hwtrust_cli_tests hwtrust_tests libhwtrust_tests
Bug: 293881506
Change-Id: Ic68273ad1c510ee4dff69f38a22b652e9ce8eba1